Implement a layer swipe tool
Created a new layer swipe bundle that provides the layer swipe map tool.
The button to activated the tool is inserted to basic toolbar when the
bundle is started. When activated, the topmost layer will be the swipe
layer and appear on left side while the rest of the layers will be on
the right side. Currenlty only non-vector layers can be the swipe layer.

When there are layers added to or removed from the map, the tool will
update the swipe layer accordingly.