Add a timer to delay swipe tool no raster alert
When switching the background map, multiple events including
remove, add and re-arrange will be triggered in order. The remove
layer event causes the NO_RASTER alert to be shown when the
background map layer itself swipe layer. This commit solve the issue
by adding a delay to the swipe tool deactivation and no raster alert.